Created By: @王同贵
Accepted By: @王同贵
URL: https://digit-force.coding.net/p/ai-platform/d/solution/git/merge/966

+++ solution-application/src/main/java/com/digitforce/aip/facade/SceneCmdFacadeImpl.java
@@ -44,6 +44,13 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
 
     @Override
     public Result delete(SceneDeleteCmd sceneDeleteCmd) {
+        Scene scene = sceneService.getById(sceneDeleteCmd.getId());
+        if (scene == null) {
+            throw new RuntimeException("场景不存在");
+        }
+        if (scene.getStatus() == SceneStatusEnum.ONLINE) {
+            throw new RuntimeException("场景已上线，不能删除");
+        }
         sceneService.removeById(sceneDeleteCmd.getId());
         return Result.success();
     }
@@ -59,7 +66,13 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
 
     @Override
     public Result unPublish(SceneStatusCmd sceneStatusCmd) {
-        Scene scene = ConvertTool.convert(sceneStatusCmd, Scene.class);
+        Scene scene = sceneService.getById(sceneStatusCmd.getId());
+        if (scene == null) {
+            throw new RuntimeException("场景不存在");
+        }
+        if (scene.getStatus() == SceneStatusEnum.OFFLINE) {
+            throw new RuntimeException("场景已下线，不能重复下线");
+        }
         scene.setStatus(SceneStatusEnum.OFFLINE);
         scene.setUpdateUser(TenantContext.tenant().getUserAccount());
         sceneService.updateById(scene);
@@ -68,7 +81,14 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
 
     @Override
     public Result modify(SceneModifyCmd sceneModifyCmd) {
-        Scene scene = ConvertTool.convert(sceneModifyCmd, Scene.class);
+        Scene scene = sceneService.getById(sceneModifyCmd.getId());
+        if (scene == null) {
+            throw new RuntimeException("场景不存在");
+        }
+        if (scene.getStatus() == SceneStatusEnum.ONLINE) {
+            throw new RuntimeException("场景已上线，不能修改");
+        }
+        scene = ConvertTool.convert(sceneModifyCmd, Scene.class);
         scene.setUpdateUser(TenantContext.tenant().getUserAccount());
         sceneService.updateById(scene);
         return Result.success();
